Fix code review issues
